Added a simulated hover behaviour
Pointers can now hover over stuff, letting us trigger style changes and mouseenter/mouseleave events. Note that simulated mouse events are now restricted to elements that have an id, because the pointer needs to remember what it's hovering and it can't possibly store "the entire DOM element". Mutations of the pointer state in the Show template are no longer overwritten by the authoritative server, we only overwrite the things the server knows about.