Fixed a bug where pointers wouldn't check for hovers when their position was being animated
We previously called checkHover only when handling a pointerMessage (= a human moving a mouse). This prevented bots from ever hovering, and would prevent human pointers from checking their hover state when they were being controlled in code, too. We now provide an array of callbacks to the stepper, that get called each frame on each pointer. This means we're calling checkHover more frequently that is strictly necessary, but it's prooobably alright.