Simplified the getElementUnder logic (follows a07744c)
If we're looking at every element under the pointer already to exclude the pointer itself, we can just as well ignore those without an id. Just filter out the wrong candidates. That way, the first remaining element is guaranteed to be one with an id, that isn't the pointer itself.